requirements.txt
aiohttp>=3.9.0
py-clob-client>=0.16.0
web3>=6.0.0
python-dotenv>=1.0.0
websockets>=12.0
rich>=13.0.0
